pnmethod example fixed
