+++ docs/source/development/quadopt_example/belief_updates.py
@@ -6,8 +6,7 @@ from typing import Union
 import numpy as np
 
 import probnum as pn
-import probnum.linalg.linops as linops
-import probnum.randvars as rvs
+from probnum import linops, randvars
 from probnum.type import FloatArgType
 
 
@@ -16,7 +15,7 @@ def gaussian_belief_update(
     action: FloatArgType,
     observation: FloatArgType,
     noise_cov: Union[np.ndarray, linops.LinearOperator],
-) -> pn.RandomVariable:
+) -> randvars.RandomVariable:
     """Update the belief over the parameters with an observation.
 
     Parameters
@@ -48,4 +47,4 @@ def gaussian_belief_update(
     # Posterior Covariance
     S = Sigma - Sigma @ Phi @ np.linalg.solve(gram, Phi.T @ Sigma)
 
-    return rvs.Normal(mean=m, cov=S)
+    return randvars.Normal(mean=m, cov=S)
