+++ docs/source/development/quadopt_example/observation_operators.py
@@ -1,20 +1,17 @@
-"""
-Observation operators for 1D quadratic optimization.
-"""
+"""Observation operators for 1D quadratic optimization."""
 
 from typing import Callable
 
 import numpy as np
 
-from probnum import utils as _utils
+from probnum import utils
 from probnum.type import FloatArgType
 
 
 def function_evaluation(
     fun: Callable[[FloatArgType], FloatArgType], action: FloatArgType
 ) -> np.float_:
-    """
-    Observe a (noisy) function evaluation of the quadratic objective.
+    """Observe a (noisy) function evaluation of the quadratic objective.
 
     Parameters
     ----------
@@ -25,7 +22,7 @@ def function_evaluation(
     """
     observation = fun(action)
     try:
-        return _utils.as_numpy_scalar(observation, dtype=np.floating)
+        return utils.as_numpy_scalar(observation, dtype=np.floating)
     except TypeError as exc:
         raise TypeError(
             "The given argument `p` can not be cast to a `np.floating` object."
